fix clippy

+++ drivers/kdriver/src/bus/pci.rs
@@ -106,7 +106,7 @@ fn config_pci_device<C: ConfigurationAccess>(
 impl AllDevices {
     /// Enumerate PCI devices and register matching drivers.
     pub(crate) fn probe_bus_devices(&mut self) {
-        let base_vaddr = p2v((kbuild_config::PCI_ECAM_BASE as usize).into() );
+        let base_vaddr = p2v((kbuild_config::PCI_ECAM_BASE).into() );
         let mut root = {
             #[cfg(feature = "pci-mmio")]
             {
@@ -123,7 +123,7 @@ impl AllDevices {
         // PCI 32-bit MMIO space
         let mut allocator = kbuild_config::PCI_RANGES
             .get(1)
-            .map(|range| PciRangeAllocator::new(range.0 as u64, range.1 as u64));
+            .map(|range| PciRangeAllocator::new(range.0, range.1));
 
         for bus in 0..= kbuild_config::PCI_BUS_END as u8 {
             for (bdf, dev_info) in root.enumerate_bus(bus) {
